requirements.txt
fastapi
uvicorn[standard]
streamlit
chromadb
sentence-transformers
torch
transformers
git+https://github.com/openai/CLIP.git
Pillow
pytesseract
opencv-python
scikit-learn
python-docx
python-pptx
openpyxl
pandas
PyMuPDF
scipy
tqdm
